fabio.tests() passed
mainly changes in _read_header_block:

search.pattern corrected from
  end_pattern = re.compile(b"}[\r\n]")
to
  end_pattern = re.compile(b'}(\r{0,1})\n')
In this way, it only accepts }\r\n and }\n as
header end patterns. The end_pattern.search().end()
is now always pointing to \n.

It is not tried any more searching the key Size in
read_header_block, because it is probably not
even loaded. A simple search would also match other keys,
e.g. PSize_1 etc. and give wrong results. Only the
EDF_* keys are at the start of the header and can be
read.

If { is the last read byte, it cannot be recognized
if it is part of header_end pattern. This should not be
a problem if the files are written with a blocksize of
512 bytes. Parsing the header with an overlap of two
bytes per read should solve this problem generally.

Still to do:
  - Instead of searching the header_end pattern byte by
    byte read the header at once if HeaderSize is known.
    Then search it only in the last 3 bytes.
    In the best way all EDF_* keys should be read before
    searching the end of the header.
  - When everything works clean it from the now unnecessary
    commented parts.
  - Rework the variance / error access as proposed by
    Valentin.