Add support for GE files with an EDF header
This file type started coming from APS in ~2022 with a `.edf.ge5` extension.

In fabio, these were being interpreted as EDF files because of their EDF-like
header. With the changes in this PR, the file is interpreted as a GE file
instead and the frames are read in correctly.

The added logic is as follows:

1. If the file is identified as EDF, we check the file extension, and if the
file extension is `ge` plus a number, it is interpreted as GE instead.
2. We check for EDF header formats in the GE reader, and if we have a match,
we read in the header using EDF style and grab the number of frames from there.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>